refactor(import): move project-import logic to shared/services
TemplateLibraryDialog imported from features/import, violating the
feature-isolation rule. Move the project-import functions from
features/import/useImportProject.ts to shared/services/project-import/
projectImport.ts, drop the unused loadJSZip helper, and repoint all
three consumers (TemplateLibraryDialog, ProjectHubDialog,
useGitHubBumlImport).
